Changes: Added floating point drawing methods to MScalableImage.
RevBy: Daniel d'Andrada, Armin Berres

+++ src/views/mnavigationbarview.cpp
@@ -427,7 +427,7 @@ void MNavigationBarView::drawBackground(QPainter *painter, const QStyleOptionGra
 {
     //draw shadow under the actual navigationbar
     if( style()->dropShadowImage() ) {
-        style()->dropShadowImage()->draw(0, size().height(), boundingRect().width(),  style()->dropShadowImage()->pixmap()->size().height(), painter);
+        style()->dropShadowImage()->draw(0.0, size().height(), boundingRect().width(),  (qreal)style()->dropShadowImage()->pixmap()->size().height(), painter);
     }
 
     MWidgetView::drawBackground(painter, option);
